- extended documentation - refactoring of finder classes
