+++ licensescout-maven-plugin/src/main/java/org/aposin/licensescout/mojo/AbstractScanMojo.java
@@ -85,6 +85,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Specification of output types and filenames.
      * 
+     * <p>See also <a href="../licensescout-documentation/usermanual/configuration.html#output_types_and_files">Output Types and files</a>.</p>
+     * 
      * @see Output
      * @since 1.2
      */
@@ -94,6 +96,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Name of the file to read known licenses from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-licenses">Licenses</a>.<p>
+     * 
      * @since 1.1
      */
     @Parameter(property = "licensesFilename", required = false)
@@ -101,6 +105,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
 
     /**
      * Name of the file to read known providers from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-providers">Providers</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -110,6 +116,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Name of the file to read license notices from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-notices">Notices</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(property = "noticesFilename", required = false)
@@ -117,6 +125,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
 
     /**
      * Name of the file to read checked archives from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-checked-archives">Checked Archives</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -126,6 +136,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Name of the file to read license URL mappings from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-license-url-mapping">License URL Mapping</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(defaultValue = "urlmappings.csv", property = "licenseUrlMappingsFilename", required = false)
@@ -134,6 +146,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Name of the file to read license name mappings from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-license-name-mapping">License Name Mapping</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(defaultValue = "namemappings.csv", property = "licenseNameMappingsFilename", required = false)
@@ -141,6 +155,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
 
     /**
      * Name of the file to read global filter patterns from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-global-filters">Global Filters</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -151,6 +167,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
      * Name of the file to read of vendor names to filter out from.
      * This is alternative to {@link #filteredVendorNames}.
      * If both are given, the entries are merged.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-vendor-names">Vendor Names</a>.<p>
      * 
      * @since 1.1
      */
@@ -372,6 +390,8 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     /**
      * Skips the execution.
      * 
+     * <p>For examples on using this parameter in a build see <a href="../licensescout-maven-plugin/examples/controlling-execution.html">Controlling the execution</a>.<p>
+     * 
      * @since 1.3.1
      */
     @Parameter(defaultValue = "false", property = "skip", required = false)
@@ -399,12 +419,12 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     private MavenProject mavenProject;
 
     /**
-     * The Maven session.
+     * The Maven settings.
      * @since 1.4
      */
-    @Parameter(defaultValue = "${session}", readonly = true)
-    protected org.apache.maven.execution.MavenSession mavenSession;
-
+    @Parameter( defaultValue = "${settings}", readonly = true )
+    private Settings mavenSettings;
+    
     /**
      * The Maven project helper.
      */
@@ -772,7 +792,7 @@ public abstract class AbstractScanMojo extends AbstractMojo implements IReposito
     }
 
     protected Settings getSettings() {
-        return mavenSession.getSettings();
+        return mavenSettings;
     }
 
     /**
