+++ licensescout-parent/pom.xml
@@ -215,7 +215,6 @@
 				<groupId>org.apache.maven.plugin-tools</groupId>
 				<artifactId>maven-plugin-annotations</artifactId>
 				<version>${maven-plugin-annotations.version}</version>
-				<scope>provided</scope>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.maven.reporting</groupId>
@@ -494,6 +493,9 @@
 					<artifactId>maven-plugin-plugin</artifactId>
 					<version>${maven-plugin-plugin.version}</version>
 					<configuration>
+						<extractors>
+							<extractor>java-annotations</extractor>
+						</extractors>
 						<skipErrorNoDescriptorsFound>true</skipErrorNoDescriptorsFound>
 					</configuration>
 				</plugin>
