+++ licensescout-report-maven-plugin/src/main/java/org/aposin/licensescout/report/mojo/AbstractReportMojo.java
@@ -71,6 +71,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
     /**
      * Name of the file to read known licenses from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-licenses">Licenses</a>.<p>
+     * 
      * @since 1.1
      */
     @Parameter(property = "licensesFilename", required = false)
@@ -78,6 +80,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
 
     /**
      * Name of the file to read known providers from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-providers">Providers</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -87,6 +91,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
     /**
      * Name of the file to read license notices from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-notices">Notices</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(property = "noticesFilename", required = false)
@@ -94,6 +100,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
 
     /**
      * Name of the file to read checked archives from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-checked-archives">Checked Archives</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -103,6 +111,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
     /**
      * Name of the file to read license URL mappings from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-license-url-mapping">License URL Mapping</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(defaultValue = "urlmappings.csv", property = "licenseUrlMappingsFilename", required = false)
@@ -111,6 +121,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
     /**
      * Name of the file to read license name mappings from.
      * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-license-name-mapping">License Name Mapping</a>.<p>
+     * 
      * @since 1.2.6
      */
     @Parameter(defaultValue = "namemappings.csv", property = "licenseNameMappingsFilename", required = false)
@@ -118,6 +130,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
 
     /**
      * Name of the file to read global filter patterns from.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-global-filters">Global Filters</a>.<p>
      * 
      * @since 1.2.6
      */
@@ -128,6 +142,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
      * Name of the file to read of vendor names to filter out from.
      * This is alternative to {@link #filteredVendorNames}.
      * If both are given, the entries are merged.
+     * 
+     * <p>For the format of the file see <a href="../licensescout-documentation/usermanual/configuration.html#configuration-file-vendor-names">Vendor Names</a>.<p>
      * 
      * @since 1.1
      */
@@ -237,6 +253,8 @@ public abstract class AbstractReportMojo extends AbstractMavenReport implements 
 
     /**
      * Skips the execution.
+     * 
+     * <p>For examples on using this parameter in a build see <a href="../licensescout-maven-plugin/examples/controlling-execution.html">Controlling the execution</a>.<p>
      * 
      * @since 1.3.1
      */
